Don't use words as unit for slurping and barfing
Using words as the unit for slurping can give confusing results.  For
example, slurping forward with

   The [http://orgmode.org] Org-mode site

results in

   The [[http://orgmode.org][Org]]-mode site

This is somewhat inconvenient because the intended slurp here is most
likely 'Org-mode'.  However, given that the command clearly says that
it works at the level of words, this shouldn't be too surprising.

The problem is that, unless the user is displaying the full link
(e.g., using org-toggle-link-display), the entire text 'Org-mode' is
highlighted as though it's part of the link.  If the user then tried
to slurp again to get 'site', '-mode' would actually be slurped, and
it would appear like nothing happened.

Instead of words, make blocks of non-whitespace characters, referred
to as blobs, the unit for slurping and barfing.  When slurping
forward, don't consider trailing punctuation as part of a blob.