barf: Fix edge cases where point is beyond link
When a point is on a blank space beyond the link but before the next
element, org-element-context considers point to be a part of the link.
This leads to incorrect behavior (shown in the added tests) when
either barfing backward or forward.  Avoid these issues by moving
point to the beginning of the link before deleting the old link and
inserting a new one.  As a bonus, this makes some of the later
goto-char calculations unnecessary.